requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
black==21.12b0
cachetools==4.2.2
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
python-dotenv==0.19.2
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tomli==1.2.3
tornado==6.1
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
